requirements.txt
streamlit
langchain
langchain-community
langchain-core
langchain-google-genai
faiss-cpu
sentence-transformers
google-api-core